# Conflicts:
#	src/editor/app.tsx
#	src/editor/sprite/selection-controller.ts
#	src/editor/sprite/sprite-document.ts
#	src/editor/sprite/sprite-editor.tsx
#	src/editor/sprite/timeline.tsx
#	test/selection-free-transform.test.ts
#	test/selection-transform.test.ts
#	test/sprite-selection-controller.test.ts
#	vite.config.ts

+++ test/stroke-buffer.test.ts
@@ -152,7 +152,7 @@ class FakeDoc {
 		this.layer = {
 			width: W,
 			height: H,
-			data: (snap.data.data as Uint8ClampedArray).slice(),
+			data: snap.data.data.slice(),
 		};
 	}
